Attempt to only use this on embedding tests
