+++ e2e/support/config.js
@@ -248,7 +248,7 @@ const embeddingSdkComponentTestConfig = {
   video: false,
   specPattern: "e2e/test-component/scenarios/embedding-sdk/**/*.cy.spec.tsx",
   indexHtmlFile: "e2e/support/component-index.html",
-  supportFile: "e2e/support/cypress.js",
+  supportFile: "e2e/support/cypress-support-sdk.js",
 
   reporter: mainConfig.reporter,
   reporterOptions: mainConfig.reporterOptions,
